Simplify Enroll User API
PR #82 introduced an endpoint for issuing SSH user certificates. The API
assumed that the user param passed in the API should match the username
in the HTTP header passed by the authenticating proxy. Since, there's no
functionality, currently, that allows a user to issue a cert on behalf
of another user (e.g. authenticated user, Alice, trying to call
/enroll_user/bob), this part of the API is redundant and is ripe for
simplification.